+++ updates/updates.go
@@ -1,4 +1,4 @@
-package main
+package updates
 
 import (
 	"encoding/json"
